Pushing updates Day 6 and 7
I've still doing the CSS part of the first project,
and I'll probably work on the CSS on this one, too.
I already have some things in mind, but I'll be needing more practice,
experience, and a little bit of javascript to do what I have in mind.
Still, it's a good start.